reverted to text context for now
